split test files in half for more efficient run time
