+++ .travis.yml
@@ -1,15 +1,20 @@
 sudo: false  # Use container-based infrastructure
 language: python
 cache: 
-  pip: True
+  pip: true
+  directories:    
+    - $HOME/.cached_models
 python:
   - "2.7"
   #- "3.4"
 env:
   - TEST=test/test_exp_survival_model.py
+  - TEST=test/test_exp_survival_model_sim.py
   - TEST=test/test_pem_survival_model.py
-  #- TEST=test/test_pem_survival_model_varcoefs.py
+  - TEST=test/test_pem_survival_model_sim.py
   - TEST=test/test_weibull_survival_model.py
+  - TEST=test/test_weibull_survival_model_sim.py
+  #- TEST=test/test_pem_survival_model_varcoefs.py
 before_install:
   # Commands below copied from: http://conda.pydata.org/docs/travis.html
   # We do this conditionally because it saves us some downloading if the
