+++ test/test_pem_survival_model_sim.py
@@ -4,50 +4,25 @@ mpl.use('Agg')
 import survivalstan
 from stancache import stancache
 import numpy as np
-from functools import partial
 from nose.tools import ok_
+from functools import partial
 num_iter = 500
-from .test_datasets import load_test_dataset
+from .test_datasets import load_test_dataset_long, sim_test_dataset_long
 
-model_code = survivalstan.models.exp_survival_model
-make_inits = survivalstan.make_weibull_survival_model_inits
+model_code = survivalstan.models.pem_survival_model
+make_inits = None
 
-def test_model(**kwargs):
-    ''' Test survival model on test dataset
+def test_pem_model_sim(**kwargs):
+    ''' Test weibull survival model on simulated dataset
     '''
-    d = load_test_dataset()
+    dlong = sim_test_dataset_long()
     testfit = survivalstan.fit_stan_survival_model(
         model_cohort = 'test model',
         model_code = model_code,
-        df = d,
-        time_col = 't',
-        event_col = 'event',
-        formula = 'age + sex',
-        iter = num_iter,
-        chains = 2,
-        seed = 9001,
-        make_inits = make_inits,
-        FIT_FUN = stancache.cached_stan_fit,
-        **kwargs
-        )
-    ok_('fit' in testfit)
-    ok_('coefs' in testfit)
-    ok_('loo' in testfit)
-    survivalstan.utils.plot_coefs([testfit])
-    survivalstan.utils.plot_coefs([testfit], trans=np.exp)
-    return(testfit) 
-
-
-def test_null_model(**kwargs):
-    ''' Test NULL survival model on flchain dataset
-    '''
-    d = load_test_dataset()
-    testfit = survivalstan.fit_stan_survival_model(
-        model_cohort = 'test model',
-        model_code = model_code,
-        df = d,
-        time_col = 't',
-        event_col = 'event',
+        df = dlong,
+        sample_col = 'index',
+        timepoint_end_col = 'end_time',
+        event_col = 'end_failure',
         formula = '~ 1',
         iter = num_iter,
         chains = 2,
@@ -61,5 +36,34 @@ def test_null_model(**kwargs):
     ok_('loo' in testfit)
     survivalstan.utils.plot_coefs([testfit])
     survivalstan.utils.plot_coefs([testfit], trans=np.exp)
+    survivalstan.utils.plot_coefs([testfit], element='baseline')
+    return(testfit)
+
+
+def test_pem_model_sim_covar(**kwargs):
+    ''' Test weibull survival model on simulated dataset
+    '''
+    dlong = sim_test_dataset_long()
+    testfit = survivalstan.fit_stan_survival_model(
+        model_cohort = 'test model',
+        model_code = model_code,
+        df = dlong,
+        sample_col = 'index',
+        timepoint_end_col = 'end_time',
+        event_col = 'end_failure',
+        formula = '~ age + sex',
+        iter = num_iter,
+        chains = 2,
+        seed = 9001,
+        make_inits = make_inits,
+        FIT_FUN = stancache.cached_stan_fit,
+        **kwargs
+        )
+    ok_('fit' in testfit)
+    ok_('coefs' in testfit)
+    ok_('loo' in testfit)
+    survivalstan.utils.plot_coefs([testfit])
+    survivalstan.utils.plot_coefs([testfit], trans=np.exp)
+    survivalstan.utils.plot_coefs([testfit], element='baseline')
     return(testfit)
